requirements.txt
streamlit
mysql-connector-python
pandas
streamlit-authenticator
bcrypt
pyyaml
openpyxl
python-dotenv
plotly